+++ lib/srv/desktop/discovery_test.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/go-ldap/ldap/v3"
 	"github.com/gravitational/teleport/api/types"
+	"github.com/gravitational/teleport/lib/auth/windows"
 	"github.com/stretchr/testify/require"
 )
 
@@ -63,14 +64,14 @@ func TestDiscoveryLDAPFilter(t *testing.T) {
 func TestAppliesLDAPLabels(t *testing.T) {
 	l := make(map[string]string)
 	entry := ldap.NewEntry("CN=test,DC=example,DC=com", map[string][]string{
-		attrDNSHostName:       {"foo.example.com"},
-		attrName:              {"foo"},
-		attrOS:                {"Windows Server"},
-		attrOSVersion:         {"6.1"},
-		attrDistinguishedName: {"CN=foo,OU=IT,DC=goteleport,DC=com"},
-		attrCommonName:        {"foo"},
-		"bar":                 {"baz"},
-		"quux":                {""},
+		windows.AttrDNSHostName:       {"foo.example.com"},
+		windows.AttrName:              {"foo"},
+		windows.AttrOS:                {"Windows Server"},
+		windows.AttrOSVersion:         {"6.1"},
+		windows.AttrDistinguishedName: {"CN=foo,OU=IT,DC=goteleport,DC=com"},
+		windows.AttrCommonName:        {"foo"},
+		"bar":                         {"baz"},
+		"quux":                        {""},
 	})
 
 	s := &WindowsService{
@@ -105,21 +106,21 @@ func TestLabelsDomainControllers(t *testing.T) {
 		{
 			desc: "DC",
 			entry: ldap.NewEntry("CN=test,DC=example,DC=com", map[string][]string{
-				attrPrimaryGroupID: {writableDomainControllerGroupID},
+				windows.AttrPrimaryGroupID: {windows.WritableDomainControllerGroupID},
 			}),
 			assert: require.True,
 		},
 		{
 			desc: "RODC",
 			entry: ldap.NewEntry("CN=test,DC=example,DC=com", map[string][]string{
-				attrPrimaryGroupID: {readOnlyDomainControllerGroupID},
+				windows.AttrPrimaryGroupID: {windows.ReadOnlyDomainControllerGroupID},
 			}),
 			assert: require.True,
 		},
 		{
 			desc: "computer",
 			entry: ldap.NewEntry("CN=test,DC=example,DC=com", map[string][]string{
-				attrPrimaryGroupID: {"515"},
+				windows.AttrPrimaryGroupID: {"515"},
 			}),
 			assert: require.False,
 		},
